+++ R/neighbor_score.R
@@ -1,40 +1,44 @@
 #' Neighbor Agreement Score
 #'
-#' Computes local label agreement.
+#' Computes local neighborhood
+#' agreement scores.
 #'
 #' @param object Seurat object
-#' @param reduction Dimensional reduction method
+#' @param reduction Reduction method
 #' @param dims PCA dimensions
 #' @param k Number of neighbors
 #' @param label_column Metadata label column
 #'
 #' @return Numeric vector
 #'
+#' @examples
+#' NULL
+#'
 #' @export
 
 neighbor_score <- function(
-    object,
-    reduction = "pca",
-    dims = NULL,
-    k = 10,
-    label_column = "predicted_label"
+  object,
+  reduction = "pca",
+  dims = NULL,
+  k = 10,
+  label_column = "predicted_label"
 ) {
-
   embeddings <- Seurat::Embeddings(
     object,
     reduction
   )
 
   # Automatically determine dimensions
-  if(is.null(dims)) {
-
+  if (is.null(dims)) {
     max_dims <- ncol(
       embeddings
     )
 
-    dims <- 1:min(
-      10,
-      max_dims
+    dims <- seq_len(
+      min(
+        10,
+        max_dims
+      )
     )
   }
 
@@ -61,8 +65,7 @@ neighbor_score <- function(
     length(labels)
   )
 
-  for(i in seq_along(labels)) {
-
+  for (i in seq_along(labels)) {
     neighbor_labels <- labels[
       nn$nn.index[i, ]
     ]