+++ tests/testthat/test-match_labels.R
@@ -1,0 +1,16 @@
+test_that(
+  "match_labels identifies labels",
+  {
+    result <- match_labels(
+      "T_cells",
+      c(
+        "T_cell",
+        "B_cell"
+      )
+    )
+
+    expect_true(
+      is.character(result)
+    )
+  }
+)